+++ ICSSoft.STORMNET.Business/XMLDataService/XMLDataService.cs
@@ -1,5 +1,11 @@
 ﻿namespace ICSSoft.STORMNET.Business
 {
+    using ICSSoft.Services;
+    using ICSSoft.STORMNET.Business.Audit;
+    using ICSSoft.STORMNET.Exceptions;
+    using ICSSoft.STORMNET.FunctionalLanguage;
+    using ICSSoft.STORMNET.FunctionalLanguage.SQLWhere;
+    using ICSSoft.STORMNET.Security;
     using System;
     using System.Collections;
     using System.Collections.Specialized;
@@ -7,16 +13,9 @@
     using System.IO;
     using System.Linq;
     using System.Reflection;
-    using ICSSoft.Services;
-    using ICSSoft.STORMNET.Business.Audit;
-    using ICSSoft.STORMNET.Exceptions;
-    using ICSSoft.STORMNET.Security;
+    using Unity;
 
     using STORMFunction = ICSSoft.STORMNET.FunctionalLanguage.Function;
-    using FunctionalLanguage.SQLWhere;
-    using FunctionalLanguage;
-
-    using Unity;
 
     /// <summary>
     /// Сервис данных для XML.