+++ NewPlatform.Flexberry.ORM.IntegratedTests/ICSSoft.STORMNET.Business/DetailsLoadTest.cs
@@ -176,7 +176,8 @@
                         output.WriteLine(item.Value.ToString());
                     }
 
-                    throw exception.InnerException;
+                    // Пусть так.
+                    Assert.Empty(exception.InnerExceptions);
                 }
 
                 ds.OnGenerateSQLSelect -= ThreadTesting_OnGenerateSQLSelect;